requirements.txt
msgpack-python
pyyaml 